+++ src/socket/conferir.item/conferir.item.event.ts
@@ -75,6 +75,10 @@ export default class ConferirItemEvent {
         kind: 'mutation',
       }, async ({ request, emitMutation, emitListen }) => {
         const itens = this.convert(request.mutation);
+        if (itens.length === 0) {
+          emitMutation([]);
+          return;
+        }
         const inserteds = await this.repository.insert(itens);
 
         const basicEvent = new ExpedicaoMutationBasicEvent({
@@ -95,6 +99,10 @@ export default class ConferirItemEvent {
         kind: 'mutation',
       }, async ({ request, emitMutation, emitListen }) => {
         const itens = this.convert(request.mutation);
+        if (itens.length === 0) {
+          emitMutation([]);
+          return;
+        }
         await this.repository.update(itens);
 
         const basicEvent = new ExpedicaoMutationBasicEvent({
@@ -115,6 +123,10 @@ export default class ConferirItemEvent {
         kind: 'mutation',
       }, async ({ request, emitMutation, emitListen }) => {
         const itens = this.convert(request.mutation);
+        if (itens.length === 0) {
+          emitMutation([]);
+          return;
+        }
         await this.repository.delete(itens);
 
         const basicEvent = new ExpedicaoMutationBasicEvent({
@@ -137,7 +149,11 @@ export default class ConferirItemEvent {
           ...mutation,
           Item: normalizeExpedicaoItemSequenceKey(mutation.Item),
         }),
-      { eventName: 'conferir.item.mutation', requiredKeys: ['CodEmpresa', 'CodConferir', 'CodProduto'] },
+      {
+        eventName: 'conferir.item.mutation',
+        requiredKeys: ['CodEmpresa', 'CodConferir', 'CodProduto'],
+        allowEmpty: true,
+      },
     );
   }
 }
